pyproject.toml
[tool.poetry]
name = "speedtest_exporter"
version = "0.1.0"
description = ""
authors = ["Kohei Morita <mrtc0@ssrf.in>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
humanreadable = "^0.1.0"
speedtest-cli = "^2.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
